requirements.txt
pytest==6.2.2
pyOpenSSL==20.0.1